refactor(amber): collapse the single-subclass ReplayLogger hierarchy into one class
After EmptyReplayLogger was removed, the abstract base ReplayLogger had one
subclass and was never named as a type: ReplayLogManagerImpl builds it as
`new ReplayLoggerImpl()`, so the inferred type is the concrete class.

Merges the two into a single `class ReplayLogger` in ReplayLogger.scala,
dropping `extends ReplayLogger` and the now-meaningless `override`. The Impl
suffix goes with the base -- it advertised an interface that no longer exists.
The three method bodies are untouched; no behaviour change.

Also fixes a stale `@param channel` in the Scaladoc; the parameter has been
called channelId all along.
